1. 更新数字显示问题

+++ src/bean/ClassRating.java
@@ -2,12 +2,12 @@ package bean;
 
 public class ClassRating {
     public String className;
-    public float moral;
-    public float read;
-    public float wisdom;
-    public float health;
-    public float art;
-    public float practice;
+    public double moral;
+    public double read;
+    public double wisdom;
+    public double health;
+    public double art;
+    public double practice;
     public int star;//1, 2, 3, 4, 5, 6
     public String starName;
 }
